uploading audio
Did a lot of trial and error with uploading audio and connecting the js,html and css. also trying to create the buttons and get the audio to actual play when you press play or start and pause. i also changed the color for the fonts to make them stand and against the greens.